Expected integer with resize

+++ libs/ocr.py
@@ -20,7 +20,7 @@ class Screen:
 		self.im = ImageGrab.grab(self.bbox if self.bbox else None )
 		if Config.upscale_ocr:
 			s = self.im.size
-			ratio = 300 / s[0]
+			ratio = int(300 / s[0])
 			self.im = self.im.resize( (s[0] * ratio, s[1] * ratio) )
 
 		if "enhance" in kwargs.items() and kwargs["enhance"] is True:
